+++ src/main/java/com/example/jobsearch/dto/RespondentApplicantResponseDto.java
@@ -7,8 +7,8 @@ import java.time.LocalDateTime;
 @Data
 public class RespondentApplicantResponseDto {
     private Long id;
-    private int resumeId;
-    private int vacancyId;
+    private Long resumeId;
+    private Long vacancyId;
     private boolean confirmation;
     private LocalDateTime createDate;
 }
